Rename loader_remake to loader
